refactor: changes for convention
