+++ src/components/pages/ReviewPage.tsx
@@ -1,5 +1,5 @@
+/* eslint-disable import/order */
 import React from "react";
-import { useDispatch } from "react-redux";
 import { useLocation, useNavigate } from "react-router-dom";
 import styled from "styled-components";
 import { useAppDispatch } from "../../store/hooks";
