+++ src/components/reviews/review/ReviewText.tsx
@@ -1,3 +1,4 @@
+/* eslint-disable import/order */
 import React, { useState } from "react";
 import Span from "../../utils/Span";
