+++ src/components/utils/VerticalBar.tsx
@@ -1,6 +1,7 @@
+/* eslint-disable import/order */
 import React from "react";
 import theme from "../../styles/theme";
-import Span from "../utils/Span";
+import Span from "./Span";
 
 const VerticalBar = (): JSX.Element => {
   return (
