+++ src/main.tsx
@@ -1,3 +1,4 @@
+/* eslint-disable import/order */
 import React from "react";
 import ReactDOM from "react-dom/client";
 import App from "./App";
