+++ src/network/axios.instance.ts
@@ -1,3 +1,4 @@
+/* eslint-disable import/order */
 import axios from "axios";
 
 const instance = axios.create({
