+++ src/store/store.ts
@@ -1,3 +1,4 @@
+/* eslint-disable import/order */
 import { configureStore, combineReducers } from "@reduxjs/toolkit";
 import { persistReducer } from "redux-persist";
 import storage from "redux-persist/lib/storage";
