+++ src/styles/GlobalStyle.ts
@@ -1,3 +1,4 @@
+/* eslint-disable import/order */
 import { createGlobalStyle } from "styled-components";
 
 const GlobalStyle = createGlobalStyle`
