+++ src/types/dto.ts
@@ -1,6 +1,5 @@
 export interface HospitalType {
   [key: number]: string;
-
   id: number;
   name: string;
   total_score: number;
